Implement explicit primitives.
The following explicit primitives (exprims) are implemented:

- int
- int64
- float
- double
- token
- BinaryString
- ProtectedString
- Content
- SharedString

Notably, the "number" (alias for double) and "string" types are *not*
exprims.

Exprims are typed so that a property can be set to a particular type of
value that would otherwise be ambiguous. They are userdata with a type
as usual, but have no constructors. Instead, they are constructed
through the "types" library.

The usual operations are possible on exprims, but the result is the
actual primitive value, rather than an exprim of the same type. An
exprim is meant to be short-lived.

Exprims can operate with any similar type. For example, any number-like
exprim can be added with a number or any other number-like exprim.

The normal case is that an exprim type reflects between the Go type and
the actual primitive Lua type. An exprim type can reflect from both the
primitive Lua type and the constructed type, but always reflects only
to the primitive Lua type.