Remove access to double exprim.
Like `string` maps 1:1 with Lua strings, `double` should map 1:1 with
Lua numbers. This is so that getting the value of a `double` property
returns a number rather than an exprim.

Note that the double userdata still exists and can be reflected from,
but there are no means of constructing it. This may be removed later.