Remove operators from exprim types.
Exprims are better off as containers that are as shorted-lived as
possible. The user shouldn't expect to use them anywhere except next to
getting or setting a property.

The user can operate on the underlying value of an exprim by retrieving
it from the exprim's "Value" field. A Value-field was chosen as the API
because __index can't do anything other than get a value (__call can
receive arguments, which would always be unused), and the API is
somewhat familiar (ValueBases).

For identification purposes, __tostring still indicates the exprim's
type.

Once descriptors are implemented, exprims will become a secondary
feature anyway, so they don't need to be any more convenient than
necessary.