Don't cache exprim userdata types.
Types like BinaryString and SharedString have underlying types that
cannot be compared, and therefore cannot be cached. Attempting to wrap
these in a userdata causes a panic.

Exprims are a niche feature, and exprim values are not meant to be used
anywhere except next to properties. So, instead of constructing a
cached exprim userdata, a new one is created directly. This means that
exprims cannot be properly used as table keys, but they aren't meant to
be used that way anyway.

Closes #39.