Implement clipboard source.
Format gains the MediaTypes field. This field lists a number of media
types to be utilized by sources.

Only implements source for clipboard on Windows. An implementation for
Macs will require contribution. Linux is also possible, but low
priority due to lack of support by Roblox. Ideally, this would all be
handled by a 3rd-party package, but there seems to be nothing that
handles multiple formats.

The internal clipboard is implemented as a package with Read, Write, and
Clear functions.