Implement instance attributes.
Add RBXAttr format that encodes serialized attributes. Similar to the
RBX formats, the primary purpose of this format is to convert between
rbxattr values and rbxmk values. This introduces a dependency on
github.com/robloxapi/rbxattr.

Add AttrConfig type that configures which property attributes are
applied to, defaulting to AttributesSerialize. Like RootDesc, this can
be applied to instances on an individual basis, and shares the same
inheritance behavior.

Additionally, AttrConfig can be used to select the format version, in
case that becomes necessary in the future.

Add sym.AttrConfig and sym.RawAttrConfig to access the AttrConfig on an
instance.

Add rbxmk.globalAttrConfig to apply an AttrConfig to all instances
globally. This slightly refactors indexing the rbxmk library by
handling fields more generally.

Add attribute methods to Instance reflector.
- Notably, these methods cannot be added to rtypes.Instance, because
  they require the formats package, which depends on rtypes. To apply
  attributes to rtypes.Instance using formats.RBXAttr requires a third
  party, which is done from reflect.Instance.
- Adds GetAttribute, GetAttributes, SetAttribute, and additionally
  SetAttributes, which is not present in Roblox's API (yet?).

In the documentation, the inheritance behavior of certain members is
moved to its own section and generalized. This applies to the sym.Desc
and sym.AttrConfig members.