Fix certain non-exprim properties returning empty userdata.
Condition of pushPropertyTo was that any type-like types were assumed to
exprims. The string and double types are type-like, but are not
exprims. Because they have no registered metatables, but are still
passed to UserdataOf, they are incorrectly pushed as a userdata with no
metatable.

This commit changes the condition of pushing as a userdata from checking
for a type-like, to checking for the Exprim flag on the type's
reflector.