Fix string metatables.
The string metatable in gopher-lua was set using a
self-referencing __index. Because __index was included in the string
library, it was filtered by the library filter.

v0.1.3 of gopher-lua uses a non-self-referencing string metatable, so
this problem is fixed. A test has been added to detect regressions in
this area.