Implement format options.
The places where a format string is specified is now a FormatSelector
type. FormatSelector can be reflected from a string or a table. The
string determines the format as before. As a table, the format is
specified by the Format field. The table may have additional fields
that further configure the format.

Format has an Options field that specifies the available options of the
format. This maps an option name to a value type. When reflecting a
FormatSelector, only the specified options of the specified types can
be reflected. All format options are optional, and the format must be
able to handle no options being specified.

The `file` source selects the format from the file extension as usual,
but specifies no options in this case.

FormatOptions becomes an interface with a ValueOf method that receives a
Value from a field. Several helper functions are added to the format
package to retrieve common types. FormatSelector implements
FormatOptions.

CanDecode receives a FormatOptions, so that the result can depend on the
configuration.