Implement CSRF token validation.
If a response has a 403 status, and has a X-Csrf-Token header, then the
value of the token is stored. Subsequent requests to the same host will
include the stored token. The request is automatically retried, but
only once.

Adds a Client type to wrap around http.Client. A Client field is added
to World for storing state regarding HTTP requests.