Remove readSource and writeSource.
These were originally meant for low-level access to a source, without
the need for an additional library. It turns out that each source
requires a somewhat more specialized API, so it ended up that the
library became the main point of access, while read/writeSource is
secondary.

The replacement for these functions is to encode or decode in the
desired format, then use the result with the source's library,
using "bin" as the format.

This change makes the sources.md document rather bare, so documentation
of source libraries has been moved from libraries.md to here.