Refactor cookies.
Add Cookie type, which holds information about a cookie. For security,
the value of the cookie cannot be accessed.

rbxmk.cookiesFrom retrieves cookies from a known location. Currently
implements no locations, but this will be expanded in the future.

rbxmk.newCookie creates a cookie from scratch.

The Cookies type is just a table of Cookie values, added as a
convenience.

Add Cookies field to HTTPOptions and HTTPResponse. These handle the
Cookie and Set-Cookie headers, and ensures that they are well-formed.

RBXAssetOptions.Cookies is now of the Cookies type.

Add several methods to internal HTTPHeaders for getting and setting
cookies.