Prepare for implementation of file access limits.
All accesses to the file system are replaced with the sfs package. sfs
wraps various functions to check whether a given path can be accessed.

World has an FS field, which is an sfs.FS. To prevent scripts from
escaping permitted areas, all access to the file system within the
world should be done through this FS.

The access logic currently allows anything unconditionally, but this
will be implemented properly later.