Fix userdata caching.
Immutable types like Vector3, which are expected to produce different
userdata for equal values, would be incorrectly cached. Some might
consider this behavior desirable, but it differs from Roblox's
implementation, and I don't want it to become relied upon.

This is implemented with the PushPtrTypeTo function, a companion to
PushTypeTo. This should be used for mutable types that can be used as
table keys, like Instance. Immutable types should continue to use
PushTypeTo.

The motivation for using PushPtrTypeTo is for producing consistent
userdata for a value. The motivation for using PushTypeTo is for
improved push/pull speed (the global userdata cache doesn't have to be
consulted).