Fix race condition with GC and userdata cache.
This change addresses a flaw in the userdata cache, discovered while
reviewing the implementation of go4.org/intern.

The GC can run during World.UserDataOf, invoking the finalizer of the
userdata that was just acquired. Once UserDataOf unlocked, the
finalizer will finish, incorrectly deleting the userdata from the
cache.

This change adds a flag that notifies the finalizer to reset itself when
the userdata is "resurrected" (inspired by go4.org/intern).