Refactor command documentation to use fragments.
Adds the fragments package, which implements localization. Any number of
languages can be selectively embedded by building with lang_* tags. A
language may then be selected by setting the Language variable to one
of these embedded languages. en-us is embedded and selected by default,
but can be disabled by including the lang_no_default tag.

For translation, en-us is the canonical version of documentation. The
content of each folder consists of a number of "fragments". Each
fragment is an embedded file that contains a portion of text referred
to by the program. The `rbxmk dump fragments` command will list the
fragments referred to by the program. Each reference is a path that
drills through the embedded directory, and into a file.

The format for each file is Markdown. Each heading delimits a "section",
which can be referred to by a drill path. For example,

    commands/run.md/Flags/debug

This path drills into the "commands" directory, then the "run.md" file,
then into the "Flags" section, and finally the "debug" section. This
reference resolves into the content of the debug section.

The commands directory contains a file for each command, and several
additional files:
- main.md: Contains the usage string of the program.
- world_flags.md: Contains descriptions of common flags involved in
  initializing a Lua environment.

Each command file has the following sections:
- Summary: A short summary of the command, displayed when the command is
  listed.
- Arguments: A code block summarizing the arguments of the command.
- Description: A long and thorough description of the command.
- Flags: Descriptions of each flag. Each subsection is the name of the
  flag without leading dashes.

Not every command uses every section, so such sections may be omitted.

In general, the content of fragments are expected render in the
terminal, HTML, and raw Markdown. Sections referred to by the program
will usually be rendered in isolation, and so must be written with that
in mind.

Fragment references are verified when the program runs. If a reference
is missing, the program will panic.