Fix error when converting BinaryString/SharedString to userdata.
UserDataOf expects the given Value to be comparable. BinaryString and
SharedString are not comparable, so when these are passed to
UserDataOf, a runtime error occurs.

Several solutions were considered:
- Add AsComparable field to Reflector; a function that converts a value
  of the reflector's type to an equivalent, but comparable, value.
- Reuse types.Aliaser interface instead of AsComparable.
- Use string instead of []byte for underlying types.

AsComparable was preferred over Aliaser, because the required behavior
is too specific for what Aliaser is meant to provide.

AsComparable was ruled out, because it must return a unique type
(i.e. not types.String). Otherwise, an equivalent cached String will
conflict with this value. Some kind of internal unique type could be
created, but then we'd have to deal with making sure it's properly
converted back to the correct type as necessary.

The most simple solution, and what was chosen, is to just use a string
instead of []byte for the underlying type of BinaryString and
SharedString. Most conversions still work, and the type can be
compared. The extraneous behaviors of []byte, like mutability or the
implication of "raw binaryness", don't seem to be necessary in
practice.

Bumps github.com/robloxapi/types to v0.2.0.