Fix fragment references.
embed.FS paths are case-sensitive. For whatever reason, git committed
the upper-case fragment files as lower-case, causing the fragment
references to fail to resolve. Because my local copy is still
upper-case, the fragment references still work here. This, in addition
to having no test, managed to slip under my radar.

This commit changes the fragment references to match the file names as
committed. I've no interest in trying to commit the original upper-case
files, or dealing with the consequences, so fragment file paths will
now just be lower-case from now on.