Implement Optional type.
Within descriptors, the Optional type is represented by a "?" suffix on
the type name. The API dump already has examples of this, though none
of the uses are for properties.

Unfortunately, the API dump name uses "CoordinateFrame?" for optional
CFrames. Currently this isn't a problem, as the only use of the type is
on a function rather than a property. However, it will become a problem
if this is added as a type in the future. This could be resolved by
requesting that the type be renamed, and ensuring that the naming
between a type and its optional equivalent are consistent.

If it's too late, then rbxmk will have to implement some sort of alias
system for reflector names to match with any number of type descriptor
names. Or, hard-code CoordinateFrame as an exception (eww). Notably,
CoordinateFrame is already handled by the rbx formats.

An additional change could be to add a case for Optional to ConvertFrom
implementations. The effect of this is that non-optional properties
could be set from optional exprims more easily. However, this would
involve updating ConvertFrom for many types, and implementing
ConvertFrom for many more, so it doesn't really seem worth it.

If the Optional reflector is registered, the types library will contain
the "some" and "none" functions instead of the usual "Optional" exprim
function.