Fix encoding of service status on instances.
The rbxlx format does not encode whether an instance is a service.
Because of this, when converting from rbxlx to rbxl, instances normally
marked as services are not, causing Studio to load the file
incorrectly.

When an instance is converted for encoding or decoding while a
descriptor is set, IsService will be set according to the class
descriptor instead of the original instance. If no descriptor is set,
then the original instance is used as before.