Improve doc command.
The frag mode is now the default. List mode is implemented by adding
the --list flag.

Uses templates to render dynamic input. Added variants of fragment
functions that allow template data or functions to be passed.

Trying out a system of "message" fragments, where any form of
communication, such as an error or other inline message, is moved out
to fragments.