Drop markdown in favor of HTML.
Using markdown as the intermediate format for documentation prevents
proper support for templates.

In order to drill into a fragment correctly, the markdown needs to be
parsed for headings, with the assurance that heading-like syntax within
areas like codeblocks are not mistakenly detected.

The problem is that goldmark, and pretty much every other markdown
parser, does not include a serializable AST. That is, it is not
possible to parse the markdown only for delimiting sections, then
convert directly back into the original content. The AST *can* be
rendered back to markdown, but this process is lossy. The main issue is
that the formatting of template directives are not preserved.

I've decided that markdown is more trouble than it's worth. So, HTML
shall be the intermediate format. This has several distinct
advantages:
- The html package is mostly lossless. At worst, poorly-formatted
  content is corrected. More importantly, template directives are
  preserved.
- HTML is very easy to manipulate. The goquery package makes this
  trivial.
- Templates have first-class support for HTML.

All fragments are converted to HTML. Instead of delimiting by headings,
nested <section> elements are used. Each element has a "data-name"
attribute to indicate the name. When rendering, headings will be
inserted. A notable advantage with this method is that sections have no
levels to worry about.

The concept of orphaned sections has been removed. They can be added
back in if this turns out to be needed.

The terminal renderer has been stripped down to only render the HTML
directly. This will be implemented properly later.