Implement interactive command mode.
If the executable is run from Windows Explorer, then the program will
enter a mode that allows commands to be used interactively.

NOTE: Cobra commands are currently not designed to be run more than
once. Because of this, the behavior is locked behind
the "interactive_commands" build tag. Progress on a solution can be
tracked here: https://github.com/spf13/cobra/issues/1419