Register commands and flags with associated dumps.
Commands and flags are registered with an associated dump struct when
constructed. This lays the groundwork for including commands in the
dump output. The dump will be able to get the dump struct associated
with the command or flag, gaining access to the original unresolved
fragment reference and additional information.

To reduce redundancy, registration involves filling in specified
information. For example, when a dump struct specifies a fragment
reference, registration will automatically resolve the reference to the
appropriate field in the associated command or flag.

Removes the DocumentCommands function. Command fragrefs are now
specified manually.