Refactor library merging.
ImportedAs field is changed to Import, which is a list of path elements.
Each element drills into the global environment.

Libraries are now applied to an environment using a simple recursive
merge strategy, where tables are merged, and non-tables are
overwritten.

The result is that libraries are loaded exactly the same as before, but
the logic is more consistent.