Add environment hook.
During initialization of a World, an optional environment hook is called
whenever the environment is modified. It receives an "event" that
describes the location of the changed value in the environment, and a
description of the value. The description is implemented as a path
pointing to an object within a dump.Root structure.